lint: fix the RUF findings that change code
The RUF rules enabled in the previous commit surfaced 327 sites beyond the
dead directives. Each class was read before it was fixed:

* RUF043 (144): `pytest.raises(match=...)` compiles its argument as a regex,
  so `match="spec.template.spec"` also accepted "specXtemplateYspec". 87
  sites whose only metacharacter was a literal `.` now use `re.escape(...)`
  and assert the exact message; 57 that use regex operators (`.*`, `|`, `\n`,
  `x{20}`) are raw strings, so the pattern is deliberate on its face. Every
  touched module was re-run (4,846 tests) — no expectation had been relying on
  the loose match.
* RUF059 (72): unpacked names that were never read are `_`-prefixed, so a
  reader does not hunt for the use. One of the eight production sites was
  worse than an unused name: `_cleanup_stale_webhooks` in the helm-installer
  Lambda spawned a throwaway `helm --kubeconfig` subprocess "for the env" and
  discarded every result, one wasted process per cleanup. The call is gone,
  along with the three test patches that existed only to neutralize it. The
  `should_respawn` docstring claimed its reason token "lands in the
  child-lifecycle audit event"; no caller records it, so the docstring now
  says what it is for (asserting the policy row in tests).
* RUF024 (1): `dict.fromkeys(categories, [])` gave every key the same list
  object — a real bug in a live-validation test fixture, now a comprehension.
* RUF012 (2): the two production mutable class attributes are `ClassVar`.
* RUF005/RUF015/RUF046/RUF007/RUF010/RUF022/RUF036: splats instead of list
  concatenation, `next(iter(...))` instead of `list(...)[0]`, dropped
  `int(round(...))` double casts, `itertools.pairwise`, explicit `!s`
  conversion flags, sorted `__all__`, `None` last in unions.

Also the tail of the noqa cleanup for these 86 files, which carried both a
dead directive and one of the fixes above.

Verified: ruff check/format clean repo-wide; mypy clean on every CI target
(gco/cli/gco_mcp/scripts/.github/dockerfiles, stacks+app, each Lambda dir);
and the test modules covering every touched production file pass.

+++ lambda/traffic-dial-controller/handler.py
@@ -145,7 +145,7 @@ def list_endpoint_groups(ga_client: Any, listener_arn: str) -> dict[str, dict[st
                 continue
             groups[str(region)] = {
                 "arn": str(arn),
-                "traffic_dial": int(round(float(group.get("TrafficDialPercentage", 100.0)))),
+                "traffic_dial": round(float(group.get("TrafficDialPercentage", 100.0))),
             }
         token = response.get("NextToken")
         if not token:
@@ -223,7 +223,7 @@ def healthy_percent(
         if not values:
             return None
         return 100.0 * sum(values) / len(values)
-    except Exception as exc:  # noqa: BLE001 - missing telemetry means "hold"
+    except Exception as exc:  # missing telemetry means "hold"
         logger.warning("Health signal unavailable for %s: %s", region, exc)
         return None
 
@@ -232,7 +232,7 @@ def target_dial(healthy: float, min_dial: int, full_health: int) -> int:
     """Map a healthy percent to a target dial percentage."""
     if healthy >= full_health:
         return 100
-    return max(min_dial, int(round(healthy)))
+    return max(min_dial, round(healthy))
 
 
 def step_limit(current: int, target: int, max_step: int) -> int:
@@ -326,7 +326,7 @@ def publish_metrics(cloudwatch_client: Any, decisions: list[dict[str, Any]]) -> 
                 Namespace=DIAL_METRIC_NAMESPACE,
                 MetricData=metric_data[index : index + METRIC_BATCH_SIZE],
             )
-    except Exception as exc:  # noqa: BLE001 - metrics are advisory
+    except Exception as exc:  # metrics are advisory
         logger.warning("Failed to publish traffic-dial metrics: %s", exc)
 
 
@@ -339,7 +339,7 @@ def store_state(ssm_client: Any, project_name: str, state: dict[str, Any]) -> No
             Type="String",
             Overwrite=True,
         )
-    except Exception as exc:  # noqa: BLE001 - state is advisory
+    except Exception as exc:  # state is advisory
         logger.warning("Failed to store traffic-dial state: %s", exc)
